+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Add support for object list using certain Admin SDKs (#5208)
+- Fixes source token expiration issue by acquiring new source token upon expiration.